+++ strr-platform-web/.env.example
@@ -47,3 +47,4 @@ PLAYWRIGHT_TEST_BCSC_PREMIUM_ACCOUNT_NAME=""
 PLAYWRIGHT_TEST_BCEID_USERNAME=""
 PLAYWRIGHT_TEST_BCEID_PASSWORD=""
 PLAYWRIGHT_TEST_BCEID_PREMIUM_ACCOUNT_NAME=""
+PLAYWRIGHT_TEST_BCEID_OTP_SECRET=""
