+++ strr-platform-web/package.json
@@ -39,7 +39,8 @@
     "playwright-core": "^1.49.1",
     "sass": "^1.77.6",
     "typescript": "^5.5.3",
-    "vitest": "^1.6.0"
+    "vitest": "^1.6.0",
+    "otpauth": "^9.3.6"
   },
   "dependencies": {
     "@daxiom/nuxt-core-layer-test": "^0.0.20",
