+++ strr-platform-web/tests/e2e/test-utils/getters.ts
@@ -1,0 +1,3 @@
+import { type Page } from '@playwright/test'
+
+export const getH2 = (page: Page) => page.getByTestId('h2').first()
